WebApplication is now configured in BeforeStart
